Refactor helpers to Pathfinder
